requirements.txt
# Core
fastapi==0.95.0
uvicorn[standard]==0.21.1
python-dotenv==1.0.0
pydantic==1.10.7
python-multipart==0.0.6

# Database
sqlalchemy==2.0.9
alembic==1.11.1
psycopg2-binary==2.9.6
redis==4.5.5

# AI/ML
openai==0.27.7
python-docx==0.8.11
mermaid-markdown==0.0.8

# Utils
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0
httpx==0.24.0
